test : CommentService 의 댓글 삭제 메서드 테스트 코드 변경

+++ cotobang/src/main/java/com/example/cotobang/service/CommentService.java
@@ -48,7 +48,12 @@ public class CommentService {
     }
 
     public Comment deleteComment(Long id) {
+    public Comment deleteComment(Long id, User user) {
         Comment comment = getComment(id);
+
+        if (!user.compare(comment.getUser())) {
+            throw new NotAuthorityException(user.getName());
+        }
 
         commentRepository.delete(comment);
